feat(badge notification): quand un instructeur unfollow un dossier, on ne supprime pas les notifications 'all'

+++ app/models/dossier_notification.rb
@@ -70,9 +70,17 @@ class DossierNotification < ApplicationRecord
       .destroy_all
   end
 
-  def self.destroy_notifications_instructeur_of_dossier(instructeur, dossier)
+  def self.destroy_notifications_instructeur_of_unfollowed_dossier(instructeur, dossier)
+    instructeur_preferences = instructeur_preferences(instructeur, dossier.procedure)
+
+    notification_types_to_destroy = notification_types.keys.reject do |notification_type|
+      instructeur_preferences[notification_type] == "all"
+    end
+
+    return if notification_types_to_destroy.empty?
+
     DossierNotification
-      .where(instructeur:, dossier:)
+      .where(instructeur:, dossier:, notification_type: notification_types_to_destroy)
       .destroy_all
   end
 
@@ -236,5 +244,17 @@ def find_or_create_notification(dossier, notification_type, groupe_instructeur_i
     instructeur_id:
   ) do |notification|
     notification.display_at = display_at
+  end
+end
+
+def instructeur_preferences(instructeur, procedure)
+  if (instructeur_procedure = InstructeursProcedure.find_by(instructeur:, procedure:))
+    notification_types.keys.index_with do |notification_type|
+      instructeur_procedure.notification_preference_for(notification_type)
+    end
+  else
+    notification_types.keys.index_with do |notification_type|
+      InstructeursProcedure::DEFAULT_NOTIFICATIONS_PREFERENCES[notification_type.to_sym]
+    end
   end
 end
